Move UI REST resource to separate bundle
Add UI REST bundle to openhab-core-ui feature.
