+++ features/karaf/openhab-core/src/main/feature/feature.xml
@@ -415,6 +415,7 @@
 		<feature>openhab-core-model-item</feature>
 		<feature>openhab-core-model-sitemap</feature>
 		<bundle>mvn:org.openhab.core.bundles/org.openhab.core.ui/${project.version}</bundle>
+		<bundle>mvn:org.openhab.core.bundles/org.openhab.core.io.rest.ui/${project.version}</bundle>
 	</feature>
 
 	<feature name="openhab-core-ui-icon" version="${project.version}">
